ONLINE Phone Book
This is a network application that manages an online phone directory. In this program, the phone book's data model is saved in a file on the server's computer. The customers are able to look up a phone number or add a name+phone number to the directory. As well as, the server is able to handle multiple clients with all clients having access to the same phone book. This object is created by the server at startup and passed to the client handlers.